chore: Update negotiation form to use decimal for amount field

+++ app/services/investments_services/update_fixed_investment_by_negotiation.rb
@@ -9,11 +9,11 @@ module InvestmentsServices
     end
 
     def new_current_value
-      @investment.current_value_cents + params[:invested_value_cents].to_i
+      @investment.current_value_cents + params[:invested_value_cents].to_d
     end
 
     def new_invested_value
-      @investment.invested_value_cents + params[:invested_value_cents].to_i
+      @investment.invested_value_cents + params[:invested_value_cents].to_d
     end
 
     def new_shares_total
